refactor: move timout into its own file
