+++ crates/lib/src/settings/mod.rs
@@ -1,0 +1,1 @@
+pub mod timeout;
